Support storage zones for fetching run outputs (#89)
## Description
This change updates the python SDK to support fetching outputs from runs that are using storage zones. See https://github.com/airplanedev/js-sdk/pull/320 for the equivalent changes in the js sdk.

Resolves DTS-479

## Testing
Testing completed successfully via deploying a test task locally and verifying that it was able to get outputs from child runs in the agent-storage-enabled and agent-storage-disabled cases.